ENH: Add synchronization of markups from orig representation
Modifying the markups in the "orig" view will synchronize the resulting markups into the "pial" and "inflated" views.
Currently need to disable the undo/redo for this to work, since there are many control points for the derived markups. Scrolling the mousewheel triggers many save states, which causes a major slowdown.