BUG: Fix traceback when creating new parameter node
When the module was open and a parameter node was already loaded, a traceback could be caused due to referencing a deleted QObject.
This was caused by VTK observers still triggering after the widget was deleted.

Fixed by marking the widget as deleted when onDestroyed is called.

TODO: This is a temporary fix since onDestroyed should remove all observers.